auto/requirements.txt
allure-pytest    -     2.13.5
allure-python-commons- 2.13.5
attrs          -       23.2.0
certifi        -       2024.2.2
cffi           -       1.16.0
charset-normalizer  -  3.3.2
colorama       -       0.4.6
flake8         -       7.0.0
greenlet       -       3.0.3
h11            -       0.14.0
idna           -       3.7
iniconfig      -       2.0.0
mccabe         -       0.7.0
outcome        -       1.3.0.post0
packaging       -      24.0
pip            -       24.1.2
pluggy         -       1.5.0
psycopg2-binary  -     2.9.9
pycodestyle   -        2.11.1
pycparser     -        2.22
pyflakes      -        3.2.0
PySocks       -        1.7.1
pytest        -        8.2.0
python-dotenv -        1.0.1
requests      -        2.32.2
selenium      -        4.21.0
sniffio       -        1.3.1
sortedcontainers -     2.4.0
SQLAlchemy       -     1.4.25
trio             -     0.25.1
trio-websocket   -     0.11.1
typing_extensions -    4.11.0
urllib3      -         2.2.1
webdriver-manager  -   4.0.1
wsproto      -         1.2.0